1ra Entrega Proyecto final

+++ routes/productRouter.js
@@ -1,54 +1,61 @@
 const express = require('express')
-const products = require('../class/productsClass')
-const { Router } = express 
+const products = require('../api/productsClass')
+const user = require('../user/user')
+const { Router } = express
 const productRouter = Router()
 
-/*get productos*/
 productRouter.get('/productos', async (req, res) => {
   const allProducts = await products.getAll()
-  res.json( allProducts )
+  res.json(allProducts)
 })
 
-/*get producto segun id*/
 productRouter.get('/productos/:id', async (req, res) => {
-  const id = Number(req.params.id)
-  const product = await products.getById( id )
-  product ? res.json( product )
-    : res.status(404).send({ error: 'producto no encontrado'})
+  const id = req.params.id
+  const product = await products.getById(id)
+  product ? res.json(product)
+    : res.status(404).send({ error: 'producto no encontrado' })
 })
 
-/*post producto*/
 productRouter.post('/productos', async (req, res) => {
-  const productToAdd = req.body
-  await products.save( productToAdd )
-  res.redirect('/')
-})
-
-/*put producto*/
-productRouter.put('/productos/:id', async (req, res) => {
-  const id = Number(req.params.id)
-  const productToModify = req.body
-
-  if(await products.getById( id )){
-    let allProducts = await products.getAll()
-    allProducts[ id - 1 ] = {"id": id, ...productToModify}
-    products.saveFile( allProducts )
-    res.send({ productToModify })
+  if (user.administrador) {
+    const productToAdd = req.body
+    await products.save(productToAdd)
+    res.redirect('/')
   } else {
-    res.status(404).send({ error: 'id no valido'})
+    res.status(403).send({ error: -1, descripcion: 'ruta /productos metodo POST no autorizado' })
   }
 })
 
-/*delete producto*/
-productRouter.delete('/productos/:id', async (req, res) => {
-  const id = req.params.id
-  const productToDelete = await products.getById( id )
-
-  if (productToDelete) {
-    await products.deleteById( id )
-    res.send({ borrado: productToDelete})
+productRouter.put('/productos/:id', async (req, res) => {
+  if (user.administrador) {
+    const id = req.params.id
+    const productToModify = req.body
+    let allProducts = await products.getAll()
+    const index = allProducts.findIndex(item => item.id === id)
+    if (index !== -1) {
+      allProducts.splice(index, 1, { ...productToModify })
+      products.saveFile(allProducts)
+      res.send({ productToModify })
+    } else {
+      res.status(404).send({ error: 'id no valido' })
+    }
   } else {
-    res.status(404).send({ error: 'producto no encontrado'})
+    res.status(403).send({ error: -1, descripcion: 'ruta /productos/id metodo PUT no autorizado' })
+  }
+})
+
+productRouter.delete('/productos/:id', async (req, res) => {
+  if (user.administrador) {
+    const id = req.params.id
+    const productToDelete = await products.getById(id)
+    if (productToDelete) {
+      await products.deleteById(id)
+      res.send({ borrado: productToDelete })
+    } else {
+      res.status(404).send({ error: 'producto no encontrado' })
+    }
+  } else {
+    res.status(403).send({ error: -1, descripcion: 'ruta /productos/id metodo DELETE no autorizado' })
   }
 })
